feat: implement dashboard analytics and ETL router modules for poker data insights

+++ backend/src/api/routers/dashboard.py
@@ -223,7 +223,7 @@ async def get_profit_trend(filters: DashboardFilters, current_user: User = Depen
         hands_events = df.select([
             pl.col(sort_col).cast(pl.Utf8).alias("sort_date"),
             pl.col("date").cast(pl.Utf8).alias("display_date"),
-            pl.col("hero_net_profit_usd").alias("profit_event")
+            pl.col("hero_net_profit_usd").cast(pl.Float64).alias("profit_event")
         ])
         timeline_events.append(hands_events)
         
@@ -236,7 +236,7 @@ async def get_profit_trend(filters: DashboardFilters, current_user: User = Depen
               .otherwise(pl.lit("9999-12-31 23:59:59")).alias("sort_date"),
               
             pl.col("source_file").alias("display_date"),
-            (pl.col("prize") - pl.col("buy_in")).alias("profit_event")
+            (pl.col("prize") - pl.col("buy_in")).cast(pl.Float64).alias("profit_event")
         ])
         timeline_events.append(tourneys_events)
         
@@ -281,7 +281,7 @@ async def get_monthly_profit(filters: DashboardFilters, current_user: User = Dep
         sort_col = "data_limpa" if "data_limpa" in df.columns else "date"
         hands_events = df.select([
             pl.col(sort_col).cast(pl.Utf8).alias("sort_date"),
-            pl.col("hero_net_profit_usd").alias("profit_event")
+            pl.col("hero_net_profit_usd").cast(pl.Float64).alias("profit_event")
         ])
         timeline_events.append(hands_events)
         
@@ -290,7 +290,7 @@ async def get_monthly_profit(filters: DashboardFilters, current_user: User = Dep
             pl.when(pl.col("source_file").str.contains(r"\d{8}"))
               .then(pl.col("source_file").str.extract(r"(\d{8})").str.replace(r"(\d{4})(\d{2})(\d{2})", r"${1}-${2}-${3}"))
               .otherwise(pl.lit("9999-12-31")).alias("sort_date"),
-            (pl.col("prize") - pl.col("buy_in")).alias("profit_event")
+            (pl.col("prize") - pl.col("buy_in")).cast(pl.Float64).alias("profit_event")
         ])
         timeline_events.append(tourneys_events)
